+++ test/harness/helper.h
@@ -1,6 +1,8 @@
 #ifndef WASMVM_TEST_HELPER_DEF
 #define WASMVM_TEST_HELPER_DEF
 
-char* load_file(const char* filename, long* size);
+#include <defines.h>
+
+char* load_file(const char* filename, size_t* size);
 
 #endif
